openssl/asn1: provide methods for converting between Asn1Time and time::Tm
Openssl's APIs for this are rather minimal: while it (internally) has
code to perform the convertion between libc::tm & ASN1_TIME (and
vise-versa) it does not expose those publically.

It does, however, always store the time as an ASN.1 time string, with
either 2 or (more commonly) 4 digits for the year.

It also allows setting an ASN1_TIME via a string formatted in that way.

As a result, we add 2 free functions to perform the ASN.1 time string
to time::Tm conversion (and the inverse).